SW-3683 Stop using /api/v1/seedbank/values/all (#1506)
The seedbank API has two endpoints to list all the values of a field. One of
them lists all the values that are currently present on any accessions, and
the other lists all possible values, even if they're not used.

In early versions of the app, there was a list of seed collectors that was
maintained separately from the accession data; you could add names to the list
and they would show up as available collectors when creating a new accession.
The app thus called the "list all possible values" endpoint since it needed to
include collectors that hadn't yet been associated with any accessions.

But we moved away from that model, and the collector name is now freeform text
with a typeahead that suggests the existing collector names. There is no list
of names separate from the accession data, so the "list all possible values"
endpoint returns exactly the same results as the "list all values found on all
accessions" one.

Getting the list of collector names was the only remaining use of the "list all
possible values" endpoint. We want to remove that endpoint from the server;
switch the client over to the other one in preparation.