SW-3879 Planting Density Card and Chart Update (#1542)
Added a planting density card, which shows the total observed and target
planting density for the planting site. It also get the plantings since
the most recent observation and gives an estimate of the updated
planting density.

In addition, there were some issues with updating chart values, in
particular, when updating the data for a chart it would try to create a
new chart on the existing canvas, which is not fully released by the old
chart. Instead, keep the same chart and update its data.