SW-3914 Fix Highlight for Map
Previously, the map could get into a state where the highlight from
selection of a map entity was stuck. This was due to the select state
being cleared when the map data changed and a map popup was still open
In the code for updating callbacks on map data change, only clear the
select map feature if the popup is not open. This allows the popup
close handler to properly reset this feature state.