SW-3932 Map view for simple planting sites
For planting sites with site boundaries but without planting zones or subzones
(which we refer to as "simple" planting sites), show a map view that doesn't
mention zones and subzones.

There is currently no way to create a simple planting site from the UI; that
will be added in a later change.