SW-4432 Upgrade OpenAPI code generator (#1769)
The version of openapi-typescript we've been using doesn't have great support for
polymorphic payloads. Upgrade to the latest version, which does a better job.

With the new version, we no longer need to override the generated types for the
geometry payloads or specify the discriminator values for the search payloads;
the code generator handles both properly.