SW-4652 Only export visible inventory-by-nursery fields
CSV export of the "by nursery" inventory view was requesting the same set of
fields that are used to populate the UI, including IDs and raw numeric values.
The IDs weren't useful to users, and the raw numeric values were causing the
export to fail because CSVs need to be localized and thus can't include
non-localized fields.

Update the request construction code to only include the user-visible fields
when exporting the view to a CSV.