SW-4676 Turn the query filters hook into a session filters hook that stores the selected filters into the browser session (#2018)
- The "saved" filters are now additionally backed out to the browser session. 
- The filters are based on a particular view, so they don't transfer to other views.

The way it works is that the query will take precedence over the session. If a user navigates to a view from a URL that is sent to them, for example, it will load those filters into the view and write them to their session. If a user navigates to the view without query params, and there are filters present in their session, they will be applied to the view and written to the address bar.