Refactor the LocationSelection component within the report form, split into 3 separate locations (#2017)
This large PR just splits up the `LocationSelection` component a bit to make it a little easier to understand.

The only functional change is that the dispatches to get data needed in all the planting site location selections is within the `ProjectForm` component, so as to avoid duplicate API calls since the data is bound to the organization and not any particular planting site.