SW-4826 Viability test fixes (#2143)
We were running the validation of the number of seeds tested even when
editing an existing viability test which was triggering false failures.
Do not validate number of seeds in viability test after the test has
been created.

Do not allow for editing number of seeds in viability test after the
test has been created. We don't allow editing previously withdrawn
quantities elsewhere in the product.