SW-4643 Add additional tests for locations and accessions (#2138)
Added a test list which allows me to serialize tests and use outputs
from one as inputs to another.

This is a tradeoff because some tests will be dependent on others and
therefore running each test independently is not possible. The
alternative is to build out a lot more test infrastructure that can
pre-seed all the scenarios we need. Would like to get thoughts on this.

Added tests for the following scenarios:

- Adding an accession
- Withdrawing seeds to nursery
- Withdrawing seeds for viability test
- Creating an organization
- Creating a nursery