Create searchAndSort function which acts as a local Search API replacement that can use the same interfaces (#2242)
I did my best replicating the Search API functionality locally without sacrificing the payloads that are normally sent. This means that we don't need to change the way that the `SearchFiltersWrapper` works, even if we aren't actually using the Search API. 

I don't think this is a bulletproof implementation, and things like `Range` and `ExactOrFuzzy` aren't implemented (although the latter sounds like it would be trivial). Hopefully the test coverage is adequate to cover the most common use cases.