Fix deliverable refresh on update (#2287)
The deliverable refresh was broken when I had to modify the fetch to use deliverableId and projectId because the flow was changed to using the asyncThunk requestId for retrieving out of state. This additionally stores the deliverable at a predictable location so consumers unaware of the request ID can be updated when the deliverable is updated in state.


![2024-03-12 10.28.01.gif](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/wJnPX9wLZAkiGXEBFJxL/e601ac27-8f7b-4821-9531-58659257ef62.gif)