SW-5184 Updated withdrawal reassignments to allow selecting zones (#2504)
Updated planting zones to be updated when a withdrawal is reassigned. 

Changes:

Show planting zones in To: Subzone because subzones under different planting zones can have the same name.
![Screenshot 2024-04-10 at 5.10.30 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/NOxkRPuhAfT4F7nmdXtG/7279118b-8c71-484f-b15b-3daa59e602b5.png)


![Screenshot 2024-04-10 at 5.10.03 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/NOxkRPuhAfT4F7nmdXtG/d0b1212a-f00a-4245-9317-366d461f23ec.png)


New reassignment screen. New subzones are not populated until the zone is selected. 
![Screenshot 2024-04-10 at 5.10.47 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/NOxkRPuhAfT4F7nmdXtG/ec63f54c-e32e-4c0a-9893-686ec8ed04c3.png)

When the same zone is selected, it hides the same subzone option. 
![Screenshot 2024-04-10 at 5.11.21 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/NOxkRPuhAfT4F7nmdXtG/f90ff5a2-12c6-4bc2-83a4-9983c4b315e9.png)

When a subzone is selected, the zone is locked out, until subzone is cleared. This is because you can update the existing zone without updating subzones, and lead to confusion (especailly with name collisions).
![Screenshot 2024-04-10 at 5.11.03 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/NOxkRPuhAfT4F7nmdXtG/70654aee-ac77-44f1-a6b9-0cf65a8c9afb.png)