SW-5559 Accelerator admin deliverable categories
Add an editable list of deliverable categories for each person to the accelerator
console's "People" views.

Deliverable categories are read and updated separately from global roles in the
API, so this required adding support for the new API endpoints and calling them
from the scene code. Since both reading and writing now involve multiple API
requests, and we need the data to stay in sync between the edit and read-only
views, the API interactions are moved to `PersonProvider`, which now exposes
some callback functions to its child components.