Bring in missing PDD code (#2826)
When I originally brought the PDD code in, I missed a few commits, I did not have my local main up to date. This PR brings those missing commits into the codebase

Here are the missing commits:
https://github.com/terraware/pdd-web/pull/111
https://github.com/terraware/pdd-web/pull/112
https://github.com/terraware/pdd-web/pull/113
https://github.com/terraware/pdd-web/pull/114
https://github.com/terraware/pdd-web/pull/115
https://github.com/terraware/pdd-web/pull/116
https://github.com/terraware/pdd-web/pull/117

This PR does not contain any of the missing commits related to the preview, since that entire area will be getting an overhaul.