Update the citations display in the document producer preview (#2919)
I was unable to get the citation number to be superscript within the footnotes. This seems to be because of the fact that PagedJS renders these using the `::marker` pseudo element, which does not allow for properties such as `vertical-align: super;` or
`font-variant-position: super;` because it is displayed with `display: static` and can not be overridden. 

I also tried to see if there was a way to wrap the actual marker with a `<sup>` tag, but this also seems to be impossible since the content of the element is populated within the CSS, which does not allow for HTML tags.

Hopefully this is good enough for now and we can attempt to style it differently if we need in the future.

![SCR-20240729-ovbw.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/wJnPX9wLZAkiGXEBFJxL/e8cae39e-dd43-436c-b23b-664fcf16a68f.png)