SW-5800 Don't show "undefined" as module title (#3070)
On the deliverables list page in the accelerator console, we show a
column that has the cohort-module-specific module title (which is
usually something like "Module 1") followed by the module name. For
document deliverables whose documents were imported from PDH, there
won't necessarily be a cohort module, and thus there won't be a module
title.

Update the rendering code to show just the module name in that case, so
it doesn't show up like "undefined - GIS".