Implement "document variables" and DocumentProducerProvider (#3077)
- Implement `DocumentProducerProvider`, which should be used when
interacting with documents within the document producer. It bundles all
the necessary requests and combines the data in the ways needed for
frontend use.
- Create `useMultiSelectorProcessor`, in the same vain as
`useSelectorProcessor`, but keeps track of and combines statuses from
multiple `AsyncRequest` s.
- Within the new provider, extract the document-specific variables and
make them available to consumers, this is used to ensure that "outdated"
variables are still accessible within the document producer for sections
that have outdated variables injected into them.
- Refactor all usages of the variables / document selectors to use the
provider.
- Fix issue with document tab variable modal not showing section
instances.