SW-5884 Export observation plot information (#3223)
Add an item to the observation table view's row menus to export a CSV
file with the locations of the plots of in-progress observations.

For observations that aren't in progress, the menu item is disabled and
has a tooltip explaining why. To support that, `TableRowPopupMenu`
now supports a `tooltip` property on menu items.