SW-6165 Populate accelerator orgs with ones that contain a project with an application (#3331)
Within the Single and Edit views for a project within the accelerator console, we need to populate organizations that contain at least one project with an application. This is because a project may have filed their application, but not yet accepted as a participant. In this state, we still need to be able to modify the project lead, which is technically the "organization TF contact", which requires us to have access to the organization even though it is not directly related to the accelerator yet.

This PR modifies the `ParticipantProjectProvider` to achieve this desired result, and fixes the issue of an Accelerator Admin (with or without any relation to any particular org) not being able to modify the project lead for a project that is not directly attached to a participant or an org with an "accelerator" internal tag.