SW-6162 Fix document upgrade modal visibility (#3340)
This PR fixes the component state and some checks to make sure that the upgrade modal is only visible when a yet-to-be-completed upgrade is available for the current document.

Although this can be deployed on its own, [this](https://github.com/terraware/terraware-server/pull/2612) PR needs to be merged to QA the ticket.