SW-6716 Show observations of deleted zones/subzones
If a planting zone or subzone is deleted after an observation has
happened there, its ID no longer exists on the server side, but there's
still a record of it, which includes its name, in the planting site
history.

The server will now include names in observation results, and we can
use those to uniquely identify regions even after they've been deleted.

Update the observation views to use zone names rather than IDs and to
stop assuming that a zone or subzone in an observation result still
exists in the current version of the site.

This simplifies some code that needed to figure out the zone or subzone
name for an observation result. As part of that simplification, this
change also removes some code that figured out the zone boundary for an
observation result; the boundary was never actually used anywhere since
the map display code gets boundaries from the planting site history.