Fix recency layer (#3790)
Previously we were getting the data to draw in the map from the
observation, and if the observation didn't include all zone/sub-zones we
needed to get the not-observed zone/subzones from a base map that has
every zone/subzone, so that everything can be drawn on the map.

Now, we are getting the map data from the observation and the
plantingSiteHistory, so we have every zone/subzone included. Now there
is no need to add the missing zone/subzones to the map