SW-6632 Remove elevation from biomass details (#3810)
We never implemented the code to calculate the elevation for a
biomass observation monitoring plot; we put a placeholder value in
the UI and then lost track of needing to replace it with a real
number. Remove the field from the UI for now rather than showing an
incorrect elevation.