Fix organization navigation (#3843)
As part of the upgrade to react-router 7, they changed the behavior for relative navigation. Navigating to a non-fully qualified path could end up adding the new path onto the existing path depending on how the router was setup. 
In this case, navigating from any page in the console to the organizations page was appending `accelerator/organizations` to the current path, instead of going to `/accelerator/organizations`. This PR fixes that by adding `/` to the app path. 
All other enums in APP_PATH were already fully qualified.