SW-6792 Add deal name to documents page query (#3852)
As part of the Project Profile update, we need to link to the Documents page filtered to just the project's deal name. To support that functionality, this PR adds the deal name as a query param such that navigating to the page with the query set will filter the page, and the query gets updated whenever the Deal Name dropdown gets modified.