SW-6707 Add species match for biomass observations
Add a "Match Unrecognized Species" item to the dropdown menu on the biomass
observation details page. It uses the same modal and the same API as on the
ad-hoc and assigned observations details pages.

Since we now have the same flow in three places, move the "save" logic to a
hook to reduce code duplication.