SW-6815 Add static map to project profile (#3865)
Add a static map of the application boundary to the project profile, only rendered if the project zone figure hasn't been uploaded yet. 

If the application boundary doesn't exist, render a light map of the country boundary.

Add a label to the maps and the project zone figure to display what is being displayed.