SW-6820 Initial version of upload photos with no styling (#3884)
Initial version of upload photos with virtually no styling.

Edit page is nearly the same for now, with minor modifications to handle multiple requests on save. This page will be restyled in future stories, and is currently behind a feature flag.

This initial work is being done in order to unblock QA on some of the view Project Profile tasks (SW-6814, SW-6815).