SW-5592 Enable eslint in editors (#3886)
Delete conflicting eslint config file which caused editor linting to be disabled. 

Remove eslint manual config from package.json.

Also delete old log and tslint config files.

Attempted to remove eslintConfig from package.json as well, but this caused storybook to fail on the eslint step (perhaps because it couldn't find the eslint config). After a few hours on this, I gave up and reverted just that.

Special nod to this PR from Nick which did some of this work, and to @tommylau523 for converting the eslint.config.js to the new .mjs flat config.
https://github.com/terraware/terraware-web/pull/2783