SW-6242 SW-6840 Export all biomass CSVs (#3935)
Add an option to the biomass observation details page menu to 
download all three CSV files in a zip archive. This was originally
requested as part of SW-6242 but we ended up dropping it in favor of
individual file downloads. 

This is also groundwork for SW-6840 which will add another instance 
of exporting a batch of CSV files in zip format; the zipping logic is
in a helper function that can be reused.