SW-6840 Export observation results (#3976)
Add the ability to export the details of an observation, including
some subtotals that aren't shown in the UI at the moment. This data
can be used to generate statistics about observations beyond what
Terraware already calculates.

The export generates a zipfile containing two CSV files, one with
data rolled up per monitoring plot and the other with per-plot,
per-species data.

Unlike other CSV exports in the app, the files are generated locally
in the client code rather than downloaded from the server. This is
mostly because some of the columns in the CSVs (e.g., the link to the
plot details page) would be awkward to calculate using the search
API.