Enable jsx-no-bind linting rule, prepare to remove it in CI (#4072)
This rule currently adds over 1500 new errors. We want to start moving away from lambdas in jsx props, so enabling this rule to help us not use them moving forward. 
Also disable the rule in the github actions workflow comment, so that when we do enable linting in CI this rule doesn't keep us from seeing the rest of the issues.