SW-7016 Include unknown species in observation CSVs
By default, observation results don't include "species"-level data for plants of
unknown species, which is fine since we don't want them to appear anywhere in the
UI. However, they need to be included in exported CSV files.

Make the CSV export code fetch a version of the results that includes the unknown
species, and use that to generate the CSVs for export.