SW-7023 Stop using organization ID -1 as sentinel (#4109)
Currently, when there's no organization selected, the
OrganizationProvider sets the selected organization to a dummy one
with an organization ID of -1, which is treated as a sentinel value
in the code. But sometimes we forget to check for it and the -1
value ends up in API requests, which then sometimes cause server
errors.

* Change the provider's context to make the selected organization
  object optional; it will be undefined when there's no organization
  selected.
* Update all the places where we were checking for organization -1
  to instead check whether the selected organization object exists.
* Fix the TypeScript errors that indicated where we were using the
  organization ID without checking whether it was valid.

There may still be places where we're missing a required check, but
there should at least be fewer of them than before.