SW-7021 Don't download strings until locale is known (#4121)
If the user's locale is something other than English, only
download the strings for that locale rather than always starting
off by downloading English strings and only loading the correct
strings afterwards.

In addition to saving the user some bandwidth, this prevents
some races where we use the currently-loaded strings to
construct API requests, which turn out to fail because the
strings don't match the user's locale. SW-7036 is one such
example. This doesn't completely fix that issue but fixes one
cause of it.