SW-7055 Fix Last Report used for data (#4132)
Previously the data displayed for the most recent report on the project profile page was determined by the time the report was submitted (console) or published (funder). 
This caused issues in situations like this:
- Q1 published
- Q2 published
- Q1 re-published because of updates

In this situation, the data displayed should have been from Q2 but was instead displaying Q1's data because it was published most recently.

Change to using endDate for both console and funder portal.