SW-6844 Rename biomass CSVs; only allow zip export (#4148)
We can download a zip archive with all three of the biomass
observation details CSVs, so remove the options to download
the individual CSVs, which were only there initially because
we hadn't yet implemented zip downloads.

The files now have updated names, and the download logic is
moved to a helper function in anticipation of adding export
capability to the biomass observations list view.