Convert strings CSV files for all locales on edit
When a strings CSV file changes while the dev server is running,
we automatically regenerate its JavaScript file to avoid having to
restart the server. Unfortunately, we still have to restart the
server if strings are being added to the English CSV, because the
other locales' JavaScript files are missing those strings and
thus the TypeScript types don't match.

Change the regeneration code to always regenerate all the JS
files if any of the CSV files change. Then we can freely edit
the English strings file (or switch between git branches that have
different edits to the strings file) without restarting the dev
server.