SW-7088 Automatic translation to French and Spanish
Introduce our new LLM Autotranslate tool to the dev environment. It uses the
OpenAI API to automatically generate French and Spanish translations for strings.
README.md is updated with instructions for running it.

It detects string edits using hashcodes; these are stored in the comment fields
of the French and Spanish strings CSV files.

This initial change doesn't include any new translations; instead it adds
hashcodes to the existing translations from Phrase so they're grandfathered in.