SW-7086 Add plot number to biomass CSVs (#4211)
Include the monitoring plot number in the two exported CSVs for
biomass observations that didn't already have it. The plot number
is the same on every row in those CSVs since biomass observations
are always single-plot, but if people combine multiple CSVs from
different observations, this will help them tell which data came
from which plot.