SW-7119 Add support for ECS deployment
For prod and staging deploys, if the repo has environment variables
whose names are the tier name followed by `_ECS_CLUSTER` and
`_ECS_SERVICE` (for example, `STAGING_ECS_CLUSTER`), deploy to that
ECS cluster and service.

For now, we deploy to the EC2 instances whether or not we deploy
to ECS. This will let us test ECS deployment without disrupting the
existing prod/staging environments.