SW-7037 Hide withdrawals from deleted batches
If a seedling batch was deleted before we added the server-side
logic to detect deletion of batches with existing withdrawals,
its withdrawals were showing up in the withdrawal log with
quantities of 0 and a species name of "<deleted species>".
Update the search criteria to filter those withdrawals out
since they are misleading and we can't show anything useful
about them.