SW-7175 Fix withdrawal modal's Exceeds message missing (#4285)
The "Exceeds Remaining Quantity" was missing on the withdrawal modal
when By Weight was selected, if the original quantity was in a unit
other than Grams. This was corrected when the user selects and deselects
`Withdraw All`, but it needs to work all the time.

Set the units equal to the original quantity so that this message works
correctly.