SW-7267 Don't limit observation CSVs to 25 rows (#4363)
The search API defaults to a result count of 25, but for CSV
exports of observation data, we want to export the full data set.
Update the request payloads to set the desired count to 0, which
means no limit. Update the search request payload type declaration
to guard against anyone forgetting to set the count in the future.