SW-7088 Fail CI on missing translations (#4396)
Change the JavaScript strings file generation step in the CI workflow
to not default to English text for any strings that are missing
translations. This will cause a build failure since the list of
keys in English won't match the lists in other languages.

In dev environments, output warnings about missing translations, but
continue to default to English to avoid showing build failure errors
before autotranslate has finished generating translations.