SW-7379 Disable ability to remove TF Contacts of any kind in People List (#4475)
The isTfContact check only checked if the user's role was `=== 'Terraformation Contact'`. When internal users were added, this role was added on to for display, e.g. `Terraformation Contact - Project Lead (Isaac's Project)`, which meant that users could remove TF Contacts, as well as edit them. Additionally, the check didn't work when the language was in any language other than English. 
Update the check to see if the role includes the locale's string for TF Contact.