SW-7407 Don't request t0 data if no planting site (#4485)
A server error showed up for a -1 planting site, and was traced back to the plants dashboard's request for t0 data for survival rates. This was incorrect for 2 reasons:
1. Survival rates isn't enabled in prod right now
1. We shouldn't be requesting data when a planting site id is -1. 

Fix this so that the plants dashboard only queries t0 data when survival rate calculation is enabled, and only query t0 data when `plantingSite.id !== -1` (in more pages than just the plants dashboard.