SW-7456 Remove Add-Planting-Site button when user is missing permission (#4520)
An organization Manager went to the plants dashboard for an org that
didn't have a planting site. The dashboard displayed the Add Planting
Site button and the user tried to create a planting site but it threw a
server error because only org admins can create planting sites.

Remove the button from the dashboard if the user doesn't have permission
to add a planting site.