SW-7548 Remove unchanged metric targets from payload (#4595)
If a report exists that has been submitted, editing metric targets will not work for that year because the payload includes targets for all quarters, and the BE blocks the ability to update targets on reports that have already been submitted.

Remove metric targets that have not changed from payload.

Change targets to numbers.

Change metric target text field type to numeric.

Fix disabling of metric targets when submitted reports exist.