SW-7533 Disable report save while uploading and filter caption changes (#4609)
A user had a bunch of upload report photo requests for the same photos repeatedly within 1 or 2 seconds. The report form does not disable the save button while the photos are being uploaded, so when there are many and they are large, this can take a few seconds. The user likely clicked the save button multiple times while the photos were uploading (and then left the page which caused several to be abandoned).

Additionally each time the report form is saved, all photos captions are updated, regardless of if the user made changes to them. 

Disable the save button while photos are uploading (which also displays a spinner).
Only update photo captions for which the user made changes.