SW-7375 Use new user model for reports table users (#4618)
The reports table threw errors every time the reports had modifiedBy or submittedBy users that the current user didn't have access. The BE was updated to return these user's names in the report request (as well as to return "Terraformation Team" when the requested user is an internal user, and "Former User" when the user has been deleted).
Instead of querying for user names separately, use the updated BE response to display user names.