SW-7433 Add limit back to withdrawal history table, remove only on export (#4629)
A [previous PR](https://github.com/terraware/terraware-web/pull/4617) removed the limit from the Withdrawal History table so that users could see all rows in the table and on export. Revert this so that there is a limit on the data returned to the FE table, but update so that export returns the full list.