SW-7621 Add planting site name to ad-hoc plot CSV (#4697)
The export function for ad-hoc plot CSVs attempts to use TypeScript
types to guarantee that all the information it needs is present, but
we're using a type cast to bypass that safety check and pass in an
incomplete object. As a result, the exported CSV file for a single
ad-hoc observation didn't have a value in the "planting site name"
column.

Ideally we should pass in a fully-populated object, but as a stopgap,
add the planting site name.