SW-7579 Extend matrix view table to fill the viewport (#4718)
Our other tables in the app extend to fill the viewport of the page they are on, even if they don’t need all of the space. The matrix view table defaults to a max width of 200px for all columns. Change this to a default width when the viewport is smaller than the width of the table, and allow columns to expand to fill the width of the viewport.