SW-7453 Fix race conditions in observation links (#4716)
When opening a link (and sometimes when reloading the page) for an observation result, it would often navigate to the parent page. Specifically:
- linking to an observation redirects you to the general observations page
- linking to a zone redirects you to the observation page
- linking to a plot redirects you to the zone page

Fix this for the plot page.
Attempt to fix this for the zone and observation pages.

I was unable to reproduce this for the zone and observation pages, hence the "attempt".