SW-7694 Update permanence definition for Plots/Zones in survival rate settings (#4797)
When creating the list of permanent plots for survival rate settings, change the definition to include all of the following (previously was just numbers 1 & 2):

1. currently permanent
1. have a completed or abandoned observation
1. within that observation the plot was also permanent

This helps ensure that if a plot was previously temporary, had an observation, then was changed to permanent via a site geometry change, it will not be included incorrectly in the permanent plot list until it's had a completed observation.

Do the same thing for temporary plots (zones) in reverse.