Refactored Map Legend props (#4832)
For upcoming observation map changes, the legend will have more purpose than showing layers, markers and highlights. For example, there will be toggle options to show layers for Monitoring Plots. 

Rename the legend to be single-select, multi-select, and group-toggle, to reflect the functional behaviors of the legend.

Keep the map features names intact, but lift them out to have more general purposes. The component caller is now expected to handle the legend-feature interactions.