SW-7641 Make playwright tests run in parallel (#4752)
Now that the test dependencies have been removed, we're now able to run
playwright tests in parallel. This required some reconfiguring and a bit
more stabilizing.

Many of the changes are converting the test files to have their own
`test.describe` and a more useful `test.beforeEach` in them.
`test.list.ts` is no longer needed.

Also add a few extra ids to simplify or stabilize some tests.

Had to add organization selection to the `beforeEach` of several tests
because they now run in an indeterminate order and the user might be in
a different org now.

For now, just enable parallelization locally, as the GHA runner isn't
yet stable enough to handle multiple workers. Still working on this but
it will likely be in a different PR (assuming we're able to figure that
out).